Restructure gptfrenzy package
