+++ docs/spawn_system.md
@@ -1,18 +1,18 @@
 # Spawnable Persona System
 
 This guide explains how to package a persona so it can be loaded by
-`gptfrenzy.spawn.launch()` and used in different hosts such as Discord,
+`gptfrenzy.core.spawn.launch()` and used in different hosts such as Discord,
 Unreal or Unity. Each persona lives in its own directory with a
 `manifest.yaml` describing the entry point, capabilities and assets.
 
 ## `manifest.yaml` fields
 
-A minimal manifest created by `gptfrenzy.spawn.make_manifest()` looks
+A minimal manifest created by `gptfrenzy.core.spawn.make_manifest()` looks
 like:
 
 ```yaml
 sap_version: "0.3"
-entrypoint: gptfrenzy.spawn:launch
+entrypoint: gptfrenzy.core.spawn:launch
 assets: []
 capabilities:
   - text
@@ -22,7 +22,7 @@ license_ref: ./LICENSE_PERSONAS
 * **`sap_version`** – Version of the Spawn API. `launch()` only accepts
   `"0.3"`.
 * **`entrypoint`** – Import path to a launch function that returns a
-  `PersonaInstance`. The default is `gptfrenzy.spawn:launch`.
+  `PersonaInstance`. The default is `gptfrenzy.core.spawn:launch`.
 * **`assets`** – Optional list of extra files required by the persona
   (models, images, etc.).
 * **`capabilities`** – Flags that control which methods are exposed.
@@ -31,7 +31,7 @@ license_ref: ./LICENSE_PERSONAS
 
 ## How `launch()` works
 
-`gptfrenzy.spawn.launch(host, persona_path, **kwargs)` reads the
+`gptfrenzy.core.spawn.launch(host, persona_path, **kwargs)` reads the
 `manifest.yaml` inside `persona_path`. It verifies the `sap_version`,
 loads the `Persona` class from `persona.py` and instantiates it,
 passing `host` and any extra keywords. The resulting object is wrapped in
