+++ setup.cfg
@@ -4,6 +4,7 @@ version = 0.0.0
 
 [options]
 packages = find:
+include_package_data = true
 python_requires = >=3.10
 
 [options.entry_points]
